python-extractor/requirements.txt
fastapi
uvicorn
playwright
sqlalchemy
alembic
psycopg2-binary
celery
redis
pydantic
pydantic-settings
requests
beautifulsoup4
python-dotenv
greenlet>=3.0.0
sqladmin
itsdangerous
flower
